cvs uppdatering lagt in en setting för att ignorera tth inconsistency bytt ut subwcrev mot fusbars söta versiontemplate.py =)

+++ trunk/client/DirectoryListing.cpp
@@ -41,7 +41,7 @@ void DirectoryListing::loadFile(const string& name) {
 	string ext = Util::getFileExt(name);
 	if(Util::stricmp(ext, ".DcLst") == 0) {
 		size_t len = (size_t)::File::getSize(name);
-		if(len == -1)
+		if(len == (size_t)-1)
 			return;
 		AutoArray<u_int8_t> buf(len);
 		::File(name, ::File::READ, ::File::OPEN).read(buf, len);
